added/change: implemented the calculation of Particle Velocity and Acoustic Pressure.
calcCohAcoustPress() was reorganized so as to combine vertical, horizontal and rectangular arrays in the same code.
To make this possible, the code for all array types now uses the variable pressure2D instead of pressure1D,
making the code much easier on the eyes.
Also, all matfile output is now in complex format, which means it is no longer necessary to add the real and
imaginary parts of the output after importing it into matlab.
When calculating both particle velocity and acoustic pressure, the aacoustic pressure is obtained from the center
elements of the star pressure contributions.

Another note: rows and columns of pressure2D have been "switched". ( pressure2D[cols][rows] has become pressure2D[rows][cols])
This was done for performance reasons: this way the innermost for loop loops over the rightmost index of the array.