corrected travel time interp. at box exit in solveEikonalEq(). closes #19
bug detected by Orlando:
Emanuel, I noticed that there is a small bug in seikeq.for: ray
coordinates are linearly interpolated at the box limits but travel times
are not. Thus, travel time information is fine but at the last ray point.
I guess that the C equivalent of seikeq.for has the same problem. This is
hard to spot but generates wrong results when a timefront is to be
calculated.

Cheers

Orlando