Made Makefile more verbose. Added note for GCC bug number 7263.
To make it easier to read compiler output when building the distribution
pacakge, some spacing text has been added ( somthing like "Building
cTraceo for Linux i686".

The GCC (and specifically mingw) bug causes the message:
'warning: imaginary constants are a GCC extension [enabled by default]'
to be shown. This warning is incorrect (because imaginary constants are a
feature of C99) and cannot be disabled. See:
http://gcc.gnu.org/bugzilla/show_bug.cgi?id=7263
for more info